requirements.txt
scikit-learn
pandas
matplotlib
seaborn
joblib
streamlit
numpy
scipy 
pillow
tensorflow
